Refactor LibriSpeech Conformer RNN-T recipe
